+++ qupath-core/src/main/java/qupath/lib/images/servers/TransformedServerBuilder.java
@@ -32,6 +32,9 @@ import java.util.List;
 import qupath.lib.color.ColorDeconvolutionStains;
 import qupath.lib.images.servers.ColorTransforms.ColorTransform;
 import qupath.lib.images.servers.RotatedImageServer.Rotation;
+import qupath.lib.images.servers.transforms.BufferedImageNormalizer;
+import qupath.lib.images.servers.transforms.ColorDeconvolutionNormalizer;
+import qupath.lib.images.servers.transforms.SubtractOffsetAndScaleNormalizer;
 import qupath.lib.regions.ImageRegion;
 
 /**
@@ -208,6 +211,67 @@ public class TransformedServerBuilder {
 	}
 
 	/**
+	 * Subtract a constant offset from all channels, without clipping.
+	 * @param offsets a single offset to subtract from all channels, or an array of offsets to subtract from each channel.
+	 * @return
+	 */
+	public TransformedServerBuilder subtractOffset(double... offsets) {
+		return normalize(SubtractOffsetAndScaleNormalizer.createSubtractOffset(offsets));
+	}
+
+	/**
+	 * Subtract a constant offset from all channels, clipping the result to be &geq; 0.
+	 * @param offsets a single offset to subtract from all channels, or an array of offsets to subtract from each channel.
+	 * @return
+	 */
+	public TransformedServerBuilder subtractOffsetAndClipZero(double... offsets) {
+		return normalize(SubtractOffsetAndScaleNormalizer.createSubtractOffsetAndClipZero(offsets));
+	}
+
+	/**
+	 * Subtract a constant offset from all channels, then multiply the result by a scale factor.
+	 * @param offsets a single offset to subtract from all channels, or an array of offsets to subtract from each channel.
+	 * @param scales a single scale factor to apply to all channels, or an array of scale factors to apply to each channel.
+	 * @return
+	 */
+	public TransformedServerBuilder subtractOffsetAndScale(double[] offsets, double[] scales) {
+		return normalize(SubtractOffsetAndScaleNormalizer.create(offsets, scales));
+	}
+
+	/**
+	 * Scale all channels by a constant factor.
+	 * @param scales a single scale factor to apply to all channels, or an array of scale factors to apply to each channel.
+	 * @return
+	 */
+	public TransformedServerBuilder scaleChannels(double... scales) {
+		return normalize(SubtractOffsetAndScaleNormalizer.create(null, scales));
+	}
+
+	/**
+	 * Normalize stains using color deconvolution and reconvolution.
+	 * @param stainsInput stain vectors to apply to deconvolve the input image, which should relate to the original colors
+	 * @param stainsOutput stain vectors to apply for reconvolution, determining the output colors
+	 * @param scales optional array of scale factors to apply to each deconvolved channel.
+	 *               A scale factor of 1.0 will leave the channel unchanged, while a scale of 0.0 will suppress the channel.
+	 * @return
+	 */
+	public TransformedServerBuilder stainNormalize(ColorDeconvolutionStains stainsInput, ColorDeconvolutionStains stainsOutput, double... scales) {
+		return normalize(ColorDeconvolutionNormalizer.create(stainsInput, stainsOutput, scales));
+	}
+
+	/**
+	 * Normalize the image using the provided normalizer.
+	 * @param normalizer
+	 * @return
+	 * @ImplNote To use this method to create an image that can be added to a project, the normalizers must be JSON-serializable
+	 *           and registered under {@link ImageServers#getNormalizerFactory()}.
+	 */
+	public TransformedServerBuilder normalize(BufferedImageNormalizer normalizer) {
+		this.server = new NormalizedImageServer(server, normalizer);
+		return this;
+	}
+
+	/**
 	 * Apply color transforms to the image.
 	 *
 	 * @param transforms the transforms to apply
@@ -227,6 +291,7 @@ public class TransformedServerBuilder {
 	public TransformedServerBuilder applyColorTransforms(ColorTransform... transforms) {
 		return applyColorTransforms(Arrays.asList(transforms));
 	}
+
 	
 	/**
 	 * Get the {@link ImageServer} that applies all the requested transforms.